Inserting history into interface progress.
History can now be inserted into the page for the user to see. However, it works for both functions when you would want it to work only for one. If a solution is found for only one function- in which case .forEach may have to be used, then this can proceed.