Progress on pushing into html.
Array is pushed into html, however iterates through itself every time to actually do so, and repeats all of the previous obtained values. In order to avoid this, the splice command could fit. However, it would be more preferable to not remove anything from the array, but if splice is needed, then it is an option.